feat: email verification
When the user registers, it is needed to validate his email. An email containing a random generated number is sent to the user. He is redirected to the endpoint to put the code. If the code is correct, it will automatically validate the user, do the login and redirect to the home page.